+++ native_implementation/pluggins/flutter_audio_waveforms1.1.1.3/lib/src/util/check_samples_equality.dart
@@ -1,6 +1,8 @@
+import 'dart:typed_data';
+
 import 'package:collection/collection.dart';
 
 /// Checks for new and old samples equality to decide whether to process samples
 /// again or not when samples are updated.
-bool Function(List<double> list1, List<double> list2) checkforSamplesEquality =
-    const ListEquality<double>().equals;
+bool Function(List<int> list1, List<int> list2) checkforSamplesEquality =
+    const ListEquality<int>().equals;
