+++ native_implementation/pluggins/flutter_audio_waveforms1.1.1.3/lib/src/waveforms/rectangle_waveform/rectangle_waveform.dart
@@ -1,3 +1,5 @@
+import 'dart:typed_data';
+
 import 'package:flutter/material.dart';
 import 'package:flutter_audio_waveforms/flutter_audio_waveforms.dart';
 import 'package:flutter_audio_waveforms/src/core/audio_waveform.dart';
@@ -23,7 +25,7 @@ class RectangleWaveform extends AudioWaveform {
   // ignore: public_member_api_docs
   RectangleWaveform({
     Key? key,
-    required List<double> samples,
+    required List<int> samples,
     required double height,
     required double width,
     required Duration maxDuration,
